Feat: #19 Encaplsulation des composants thread + update des stories avec histoire
